Fix for clustered services
Check returns UNKNOWN, as the BIRD is setup to run on master-backup two node server cluster. For monitoring of clustered services it is required, that check_mk_agent on backup node do not return any data if service (BIRD) is not running.
Section <<<bird>>> is sent to output only if birdc is able to contact the daemon.